Add unit tests and improve migration
