+++ server/tests/Korga.Server.Tests/ChurchToolsSyncServiceTests.cs
@@ -1,0 +1,8 @@
+﻿using Korga.Server.ChurchTools;
+
+namespace Korga.Server.Tests;
+
+public class ChurchToolsSyncServiceTests
+{
+    ChurchToolsSyncService churchTools;
+}
